+++ orchagent/flex_counter/flex_counter_manager.cpp
@@ -44,7 +44,7 @@ const unordered_map<CounterType, string> FlexCounterManager::counter_id_field_lo
     { CounterType::ACL_COUNTER,     ACL_COUNTER_ATTR_ID_LIST },
     { CounterType::TUNNEL,          TUNNEL_COUNTER_ID_LIST },
     { CounterType::HOSTIF_TRAP,     FLOW_COUNTER_ID_LIST },
-    { CounterType::ROUTE,     FLOW_COUNTER_ID_LIST },
+    { CounterType::ROUTE,           FLOW_COUNTER_ID_LIST },
 };
 
 FlexManagerDirectory g_FlexManagerDirectory;
