+++ orchagent/muxorch.cpp
@@ -22,6 +22,7 @@
 #include "portsorch.h"
 #include "aclorch.h"
 #include "routeorch.h"
+#include "flowcounterrouteorch.h"
 #include "fdborch.h"
 
 /* Global variables */
@@ -29,6 +30,7 @@ extern Directory<Orch*> gDirectory;
 extern CrmOrch *gCrmOrch;
 extern NeighOrch *gNeighOrch;
 extern RouteOrch *gRouteOrch;
+extern FlowCounterRouteOrch *gFlowCounterRouteOrch;
 extern AclOrch *gAclOrch;
 extern PortsOrch *gPortsOrch;
 extern FdbOrch *gFdbOrch;
@@ -129,7 +131,7 @@ static sai_status_t create_route(IpPrefix &pfx, sai_object_id_t nh)
         gCrmOrch->incCrmResUsedCounter(CrmResourceType::CRM_IPV6_ROUTE);
     }
 
-    gRouteOrch->onAddMiscRouteEntry(gVirtualRouterId, pfx.getSubnet());
+    gFlowCounterRouteOrch->onAddMiscRouteEntry(gVirtualRouterId, pfx.getSubnet());
 
     SWSS_LOG_NOTICE("Created tunnel route to %s ", pfx.to_string().c_str());
     return status;
@@ -160,7 +162,7 @@ static sai_status_t remove_route(IpPrefix &pfx)
         gCrmOrch->decCrmResUsedCounter(CrmResourceType::CRM_IPV6_ROUTE);
     }
 
-    gRouteOrch->onRemoveMiscRouteEntry(gVirtualRouterId, pfx.getSubnet());
+    gFlowCounterRouteOrch->onRemoveMiscRouteEntry(gVirtualRouterId, pfx.getSubnet());
 
     SWSS_LOG_NOTICE("Removed tunnel route to %s ", pfx.to_string().c_str());
     return status;
