+++ orchagent/orchdaemon.h
@@ -11,7 +11,7 @@
 #include "intfsorch.h"
 #include "neighorch.h"
 #include "routeorch.h"
-#include "routeflowcounterorch.h"
+#include "flowcounterrouteorch.h"
 #include "nhgorch.h"
 #include "cbf/cbfnhgorch.h"
 #include "cbf/nhgmaporch.h"
