--- orchagent/routeflowcounterorch.h
@@ -1,11 +1,0 @@
-#pragma once
-
-#include "orch.h"
-
-class RouteFlowCounterOrch : public Orch
-{
-public:
-    RouteFlowCounterOrch(swss::DBConnector *db, const std::vector<std::string> &tableNames);
-    virtual ~RouteFlowCounterOrch(void);
-    void doTask(Consumer &consumer) override;
-};
