+++ orchagent/vrforch.cpp
@@ -12,6 +12,7 @@
 #include "vrforch.h"
 #include "vxlanorch.h"
 #include "routeorch.h"
+#include "flowcounterrouteorch.h"
 #include "directory.h"
 
 using namespace std;
@@ -19,9 +20,11 @@ using namespace swss;
 
 extern sai_virtual_router_api_t* sai_virtual_router_api;
 extern sai_object_id_t gSwitchId;
-extern Directory<Orch*> gDirectory;
-extern PortsOrch*       gPortsOrch;
-extern RouteOrch*       gRouteOrch;
+
+extern Directory<Orch*>      gDirectory;
+extern PortsOrch*            gPortsOrch;
+extern RouteOrch*            gRouteOrch;
+extern FlowCounterRouteOrch* gFlowCounterRouteOrch;
 
 bool VRFOrch::addOperation(const Request& request)
 {
@@ -106,7 +109,7 @@ bool VRFOrch::addOperation(const Request& request)
         vrf_table_[vrf_name].vrf_id = router_id;
         vrf_table_[vrf_name].ref_count = 0;
         vrf_id_table_[router_id] = vrf_name;
-        gRouteOrch->onAddVR(router_id);
+        gFlowCounterRouteOrch->onAddVR(router_id);
         if (vni != 0)
         {
             SWSS_LOG_INFO("VRF '%s' vni %d add", vrf_name.c_str(), vni);
@@ -179,7 +182,7 @@ bool VRFOrch::delOperation(const Request& request)
         }
     }
 
-    gRouteOrch->onRemoveVR(router_id);
+    gFlowCounterRouteOrch->onRemoveVR(router_id);
 
     vrf_table_.erase(vrf_name);
     vrf_id_table_.erase(router_id);
